+++ RELEASE_NOTES.md
@@ -3,7 +3,12 @@
 This file contains one line summaries of commits that are worthy of mentioning in release notes.
 A new header is inserted each time a *tag* is created.
 
-## v1.20.3 (currently main branch)
+## v1.20.4 (currently main branch)
+
+## v1.20.3
+
+- Java: Fix URI bug in Android Viewer sample when dropping some zips.
+- Vulkan: Fix "uninitialized texture" warnings from the Vulkan backend.
 
 ## v1.20.2